# nbankSwagger.yaml
openapi: 3.0.0
info:
  title: nbank
  contact: {}
  version: '1.0'
servers:
  - url: http://localhost:8080/api
    variables: {}
paths:
  /admin/registercustomer:
    post:
      tags:
        - admin
      summary: registerCustomer
      operationId: registerCustomer
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/registerCustomerRequest'
                - example:
                    identityNumber: '1234567890123459'
                    firstName: John
                    lastName: Doe
                    phoneNumber: '123456780'
                    email: waduh2@gmail.com
                    dob: 2004-07-20
                    address: 123 Main Street
                    occupation: Engineer
                    income: 50000
                    maritalStatus: SINGLE
                    highestEducation: BACHELORS_DEGREE
                    password: Password123
            example:
              identityNumber: '1234567890123459'
              firstName: John
              lastName: Doe
              phoneNumber: '123456780'
              email: waduh2@gmail.com
              dob: 2004-07-20
              address: 123 Main Street
              occupation: Engineer
              income: 50000
              maritalStatus: SINGLE
              highestEducation: BACHELORS_DEGREE
              password: Password123
        required: true
      responses:
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-constraintviolation1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T02:47:48.8083657
                      data: null
                      error:
                        name: Constraints Violations
                        detail:
                          - occupation must not be null
                          - address must not be null
                          - maritalStatus must not be null
                          - password must not be null
                          - email must not be null
                          - highestEducation must not be null
                          - phoneNumber must not be null
                          - identityNumber must not be null
                          - firstName must not be null
                          - dob must not be null
                          - income must not be null
              example:
                message: Request Failed
                timestamp: 2023-06-25T02:47:48.8083657
                data: null
                error:
                  name: Constraints Violations
                  detail:
                    - occupation must not be null
                    - address must not be null
                    - maritalStatus must not be null
                    - password must not be null
                    - email must not be null
                    - highestEducation must not be null
                    - phoneNumber must not be null
                    - identityNumber must not be null
                    - firstName must not be null
                    - dob must not be null
                    - income must not be null
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnAuthorized1'
                  - example:
                      timestamp: 2023-06-24T19:51:41.424+00:00
                      status: 401
                      error: Unauthorized
                      path: /api/admin/registercustomer
              example:
                timestamp: 2023-06-24T19:51:41.424+00:00
                status: 401
                error: Unauthorized
                path: /api/admin/registercustomer
      deprecated: false
      security:
        - basic: []
  /admin/registeraccount:
    post:
      tags:
        - admin
      summary: registerAccount
      operationId: registerAccount
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/registerAccountRequest'
                - example:
                    accountPassword: '123456'
                    status: ACTIVE
                    accountType: CHECKING
                    balance: 1000000
                    customerId: 620771e1-cdee-4180-839e-30c1c7bf20ec
                    depositMonth: 6
            example:
              accountPassword: '123456'
              status: ACTIVE
              accountType: CHECKING
              balance: 1000000
              customerId: 620771e1-cdee-4180-839e-30c1c7bf20ec
              depositMonth: 6
        required: true
      responses:
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-constraintviolation1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T02:52:37.28
                      data: null
                      error:
                        name: Constraints Violations
                        detail:
                          - customerId must not be null
                          - status must not be null
                          - accountPassword must not be null
                          - accountType must not be null
              example:
                message: Request Failed
                timestamp: 2023-06-25T02:52:37.28
                data: null
                error:
                  name: Constraints Violations
                  detail:
                    - customerId must not be null
                    - status must not be null
                    - accountPassword must not be null
                    - accountType must not be null
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnAuthorized1'
                  - example:
                      timestamp: 2023-06-24T19:51:41.424+00:00
                      status: 401
                      error: Unauthorized
                      path: /api/admin/registercustomer
              example:
                timestamp: 2023-06-24T19:51:41.424+00:00
                status: 401
                error: Unauthorized
                path: /api/admin/registercustomer
      deprecated: false
      security:
        - basic: []
...
  /admin/transaction/deposit:
    post:
      tags:
        - admin
      summary: Deposit
      operationId: Deposit
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/DepositRequest'
                - example:
                    accountNumber: '062503485239831'
                    money: 14000000
            example:
              accountNumber: '062503485239831'
              money: 14000000
        required: true
      responses:
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-AccountNotFound1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T02:55:12.309
                      data: null
                      error:
                        name: Account Not Found
                        detail: Account you are refering to could not be found
              example:
                message: Request Failed
                timestamp: 2023-06-25T02:55:12.309
                data: null
                error:
                  name: Account Not Found
                  detail: Account you are refering to could not be found
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnAuthorized1'
                  - example:
                      timestamp: 2023-06-24T19:51:41.424+00:00
                      status: 401
                      error: Unauthorized
                      path: /api/admin/registercustomer
              example:
                timestamp: 2023-06-24T19:51:41.424+00:00
                status: 401
                error: Unauthorized
                path: /api/admin/registercustomer
      deprecated: false
      security:
        - basic: []
  /admin/transaction/withdraw:
    post:
      tags:
        - admin
      summary: withdraw
      operationId: withdraw
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/withdrawrequest'
                - example:
                    accountNumber: '062503485239831'
                    money: 10000000
            example:
              accountNumber: '062503485239831'
              money: 10000000
        required: true
      responses:
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-AccountNotFound1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T03:06:42.838
                      data: null
                      error:
                        name: Account Not Found
                        detail: Account you are refering to could not be found
              example:
                message: Request Failed
                timestamp: 2023-06-25T03:06:42.838
                data: null
                error:
                  name: Account Not Found
                  detail: Account you are refering to could not be found
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnAuthorized1'
                  - example:
                      timestamp: 2023-06-24T19:51:41.424+00:00
                      status: 401
                      error: Unauthorized
                      path: /api/admin/registercustomer
              example:
                timestamp: 2023-06-24T19:51:41.424+00:00
                status: 401
                error: Unauthorized
                path: /api/admin/registercustomer
      deprecated: false
      security:
        - basic: []
  /admin/transaction/transfer:
    post:
      tags:
        - admin
      summary: Transfer
      operationId: Transfer
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/TransferRequest'
                - example:
                    from: '062503485395532'
                    destination: '062503485239831'
                    amount: 50000
            example:
              from: '062503485395532'
              destination: '062503485239831'
              amount: 50000
        required: true
      responses:
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-AccountFromFound1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T03:06:42.838
                      data: null
                      error:
                        name: Account Not Found
                        detail: Account you are refering to could not be found
              example:
                message: Request Failed
                timestamp: 2023-06-25T03:06:42.838
                data: null
                error:
                  name: Account Not Found
                  detail: Account you are refering to could not be found
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/UnAtuhorized1'
                  - example:
                      timestamp: 2023-06-24T19:51:41.424+00:00
                      status: 401
                      error: Unauthorized
                      path: /api/admin/registercustomer
              example:
                timestamp: 2023-06-24T19:51:41.424+00:00
                status: 401
                error: Unauthorized
                path: /api/admin/registercustomer
      deprecated: false
      security:
        - basic: []
...
  /atm/transaction/withdraw:
    post:
      tags:
        - atm
      summary: Withdraw
      operationId: Withdraw
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/WithdrawRequest1'
                - example:
                    money: 50000
            example:
              money: 50000
        required: true
      responses:
        '401':
          description: Unauthorized
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Unauthorized13'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T03:19:54.58
                      data: null
                      error:
                        name: Session Expired
                        detail: Please Login Again
              example:
                message: Request Failed
                timestamp: 2023-06-25T03:19:54.58
                data: null
                error:
                  name: Session Expired
                  detail: Please Login Again
        '400':
          description: Bad Request
          headers: {}
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BadRequest-ExceedingAmount1'
                  - example:
                      message: Request Failed
                      timestamp: 2023-06-25T03:02:56.156
                      data: null
                      error:
                        name: Exceeding Amount
                        detail: You have Passed Daily Maximum  WITHDRAW Transaction
              example:
                message: Request Failed
                timestamp: 2023-06-25T03:02:56.156
                data: null
                error:
                  name: Exceeding Amount
                  detail: You have Passed Daily Maximum  WITHDRAW Transaction
      deprecated: false
...
components:
  schemas:
    registerCustomerRequest:
      title: registerCustomerRequest
      required:
        - identityNumber
        - firstName
        - lastName
        - phoneNumber
        - email
        - dob
        - address
        - occupation
        - income
        - maritalStatus
        - highestEducation
        - password
      type: object
      properties:
        identityNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        dob:
          type: string
        address:
          type: string
        occupation:
          type: string
        income:
          type: number
        maritalStatus:
          type: string
        highestEducation:
          type: string
        password:
          type: string
      example:
        identityNumber: '1234567890123459'
        firstName: John
        lastName: Doe
        phoneNumber: '123456780'
        email: waduh2@gmail.com
        dob: 2004-07-20
        address: 123 Main Street
        occupation: Engineer
        income: 50000
        maritalStatus: SINGLE
        highestEducation: BACHELORS_DEGREE
        password: Password123
    Error:
      title: Error
      required:
        - name
        - detail
      type: object
      properties:
        name:
          type: string
        detail:
          type: array
          items:
            type: string
          description: ''
      example:
        name: Constraints Violations
        detail:
          - occupation must not be null
          - address must not be null
          - maritalStatus must not be null
          - password must not be null
          - email must not be null
          - highestEducation must not be null
          - phoneNumber must not be null
          - identityNumber must not be null
          - firstName must not be null
          - dob must not be null
          - income must not be null
    BadRequest-constraintviolation1:
      title: BadRequest-constraintviolation1
      required:
        - message
        - timestamp
        - data
        - error
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
        data:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/Error'
      example:
        message: Request Failed
        timestamp: 2023-06-25T02:47:48.8083657
        data: null
        error:
          name: Constraints Violations
          detail:
            - occupation must not be null
            - address must not be null
            - maritalStatus must not be null
            - password must not be null
            - email must not be null
            - highestEducation must not be null
            - phoneNumber must not be null
            - identityNumber must not be null
            - firstName must not be null
            - dob must not be null
            - income must not be null
    UnAuthorized1:
      title: UnAuthorized1
      required:
        - timestamp
        - status
        - error
        - path
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        error:
          type: string
        path:
          type: string
      example:
        timestamp: 2023-06-24T19:51:41.424+00:00
        status: 401
        error: Unauthorized
        path: /api/admin/registercustomer
    Error2:
      title: Error2
      required:
        - name
        - detail
      type: object
      properties:
        name:
          type: string
        detail:
          type: string
      example:
        name: Cannot Create User
        detail: Email / Identity Number Has Already Been Used
...
    registerAccountRequest:
      title: registerAccountRequest
      required:
        - accountPassword
        - status
        - accountType
        - balance
        - customerId
        - depositMonth
      type: object
      properties:
        accountPassword:
          type: string
        status:
          type: string
        accountType:
          type: string
        balance:
          type: integer
          format: int32
        customerId:
          type: string
        depositMonth:
          type: integer
          format: int32
      example:
        accountPassword: '123456'
        status: ACTIVE
        accountType: CHECKING
        balance: 1000000
        customerId: 620771e1-cdee-4180-839e-30c1c7bf20ec
        depositMonth: 6
...
    BadRequest-AccountNotFound1:
      title: BadRequest-AccountNotFound1
      required:
        - message
        - timestamp
        - data
        - error
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
        data:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/Error2'
      example:
        message: Request Failed
        timestamp: 2023-06-25T02:57:39.046
        data: null
        error:
          name: Account Not Found
          detail: Account you are refering to could not be found
    DepositRequest:
      title: DepositRequest
      required:
        - accountNumber
        - money
      type: object
      properties:
        accountNumber:
          type: string
        money:
          type: integer
          format: int32
      example:
        accountNumber: '062503485239831'
        money: 14000000
...
    withdrawrequest:
      title: withdrawrequest
      required:
        - accountNumber
        - money
      type: object
      properties:
        accountNumber:
          type: string
        money:
          type: integer
          format: int32
      example:
        accountNumber: '062503485239831'
        money: 10000000
    BadRequest-ExceedingAmount1:
      title: BadRequest-ExceedingAmount1
      required:
        - message
        - timestamp
        - data
        - error
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
        data:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/Error2'
      example:
        message: Request Failed
        timestamp: 2023-06-25T03:02:56.156
        data: null
        error:
          name: Exceeding Amount
          detail: You have Passed Daily Maximum WITHDRAW Transaction
...
    TransferRequest:
      title: TransferRequest
      required:
        - from
        - destination
        - amount
      type: object
      properties:
        from:
          type: string
        destination:
          type: string
        amount:
          type: integer
          format: int32
      example:
        from: '062503485395532'
        destination: '062503485239831'
        amount: 50000
    BadRequest-AccountFromFound1:
      title: BadRequest-AccountFromFound1
      required:
        - message
        - timestamp
        - data
        - error
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
        data:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/Error2'
      example:
        message: Request Failed
        timestamp: 2023-06-25T03:06:42.838
        data: null
        error:
          name: Account Not Found
          detail: Account you are refering to could not be found
    UnAtuhorized1:
      title: UnAtuhorized1
      required:
        - timestamp
        - status
        - error
        - path
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
          format: int32
        error:
          type: string
        path:
          type: string
      example:
        timestamp: 2023-06-24T19:51:41.424+00:00
        status: 401
        error: Unauthorized
        path: /api/admin/registercustomer
...
    Unauthorized13:
      title: Unauthorized13
      required:
        - message
        - timestamp
        - data
        - error
      type: object
      properties:
        message:
          type: string
        timestamp:
          type: string
        data:
          type: string
          nullable: true
        error:
          $ref: '#/components/schemas/Error2'
      example:
        message: Request Failed
        timestamp: 2023-06-25T03:19:54.58
        data: null
        error:
          name: Session Expired
          detail: Please Login Again
...
    WithdrawRequest1:
      title: WithdrawRequest1
      required:
        - money
      type: object
      properties:
        money:
          type: integer
          format: int32
      example:
        money: 50000
...
  securitySchemes:
    basic:
      type: http
      scheme: basic
    bearer:
      type: http
      scheme: bearer
security: []
...